requirements-rag.txt
requests==2.32.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.8.2
numpy==1.26.4
